Fix problem when using minfeaturesize with a LABEL_NO_CLIP (#2758)
git-svn-id: http://svn.osgeo.org/mapserver/branches/branch-5-4@8971 7532c77e-422f-0410-93f4-f0b67bdd69e2

+++ mapprimitive.c
@@ -1066,7 +1066,7 @@ static int getPolygonCentroid(shapeObj *p, pointObj *lp, double *miny, double *m
 /*
 ** Find a label point in a polygon.
 */
-int msPolygonLabelPoint(shapeObj *p, pointObj *lp, int min_dimension)
+int msPolygonLabelPoint(shapeObj *p, pointObj *lp, double min_dimension)
 {
   double slope;
   pointObj *point1=NULL, *point2=NULL, cp;
